requirements.txt
matplotlib
Pillow
PyPDF2
pdf2image
numpy
tkinter